+++ src/qicklab/__init__.py
@@ -1,18 +1,5 @@
-from .analysis.analysis_helper_TLS_comprehensive import *
-from .analysis.data_processing import *
-# from .analysis.fitting import *
-from .analysis.h5_data import *
-from .analysis.plotting import *
-
-# from .config.expt_config import *
-# from .config.round_robin_config import *
-# from .config.system_config import *
-
-# from .experiments.measurements import *
-
-from .hardware.build_state import all_qubit_state
-# from .hardware.build_task import *
-# from .hardware.qick_programs import *
-
-from .utils.file_helpers import *
-
+from .analysis import *
+from .config import *
+# from .experiments import *
+from .hardware import all_qubit_state
+from .utils import *
