+++ src/qicklab/analysis/fitting.py
@@ -21,31 +21,14 @@ Usage Example:
 
 import os
 import datetime
+
 import numpy as np
 import matplotlib.pyplot as plt
+
 from scipy.optimize import curve_fit
-from ..utils.file_helpers import create_folder_if_not_exists
 
-
-def exponential(x, a, b, c, d):
-    """
-    Exponential function for curve fitting.
-
-    The model is defined as:
-        y = a * exp(- (x - b) / c) + d
-
-    Parameters:
-        x (array_like): Independent variable.
-        a (float): Amplitude.
-        b (float): Time shift.
-        c (float): Decay constant (must be > 0).
-        d (float): Baseline offset.
-
-    Returns:
-        array_like: Calculated exponential function values.
-    """
-    return a * np.exp(- (x - b) / c) + d
-
+from ..utils.file_utils import create_folder_if_not_exists
+from  .fit_functions import exponential, lorentzian, allan_deviation_model
 
 def t1_fit(I, Q, delay_times=None, signal='None', return_everything=False):
     """
@@ -125,27 +108,6 @@ def t1_fit(I, Q, delay_times=None, signal='None', return_everything=False):
     else:
         return T1_est, T1_err
 
-
-def lorentzian(f, f0, gamma, A, B):
-    """
-    Lorentzian function used for curve fitting.
-
-    The model is defined as:
-        L(f) = A * gamma^2 / ((f - f0)^2 + gamma^2) + B
-
-    Parameters:
-        f (array_like): Frequency values.
-        f0 (float): Center frequency.
-        gamma (float): Half-width at half-maximum (HWHM).
-        A (float): Amplitude scaling factor.
-        B (float): Baseline offset.
-
-    Returns:
-        array_like: Calculated Lorentzian function values.
-    """
-    return A * gamma ** 2 / ((f - f0) ** 2 + gamma ** 2) + B
-
-
 def max_offset_difference_with_x(x_values, y_values, offset):
     """
     Find the x-value corresponding to the maximum average absolute difference from a given offset.
@@ -175,7 +137,6 @@ def max_offset_difference_with_x(x_values, y_values, offset):
             corresponding_x = x_values[i + 1]
 
     return corresponding_x, max_average_difference
-
 
 def fit_lorenzian(I, Q, freqs, metric_freq, signal='None', verbose=False):
     """
@@ -274,7 +235,6 @@ def fit_lorenzian(I, Q, freqs, metric_freq, signal='None', verbose=False):
             print("Error during Lorentzian fit:", e)
         return None, None, None, None, None
 
-
 def get_lorentzian_fits(I, Q, freqs, verbose=False):
     """
     Perform Lorentzian fits on I and Q data based on frequency values and choose the best fit.
@@ -304,13 +264,3 @@ def get_lorentzian_fits(I, Q, freqs, verbose=False):
 
     return I_fit, Q_fit, largest_amp_curve_mean, largest_amp_curve_fwhm, fit_err
 
-def allan_deviation_model(tau, h0, h_m1, h_m2, A, tau0):
-    # White noise term: ∝ τ^(-1/2)
-    term_white = np.sqrt(h0 / 2) * tau**(-0.5)
-    # Flicker noise (1/ƒ noise) term: constant with τ
-    term_flicker = np.sqrt(2 * np.log(2) * h_m1)
-    # Random walk noise term: ∝ τ^(1/2)
-    term_randomwalk = np.sqrt((4 * np.pi**2 / 6) * h_m2) * tau**0.5
-    # Lorentzian noise term (exponentially correlated noise)
-    term_lorentz = np.sqrt(A * tau0 / tau * (4 * np.exp(-tau/tau0) - np.exp(-2*tau/tau0) - 3 + 2*tau/tau0))
-    return term_white + term_flicker + term_randomwalk + term_lorentz
